requirements.txt
requests>=2.31
python-dotenv>=1.0
pytest>=8.0
responses>=0.25
